% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_auth.R
\name{s3_client_setup}
\alias{s3_client_setup}
\title{s3 create client}
\usage{
s3_client_setup(
  key = NA,
  secret = NA,
  main_folder = NA,
  bucket = NA,
  endpoint = NA,
  region = NA,
  use_config_files = FALSE
)
}
\arguments{
\item{key}{(chr) S3 key}

\item{secret}{(chr) s3 secret}

\item{main_folder}{(chr) folder added behind object}

\item{bucket}{(chr) s3 bucket}

\item{endpoint}{(chr) endpoint nécessaire pour accéder à S3}

\item{region}{(chr) s3 region, OVH == ""}

\item{use_config_files}{(log) TRUE if use with files config files}
}
\value{
paws variable
}
\description{
s3 create client
}
\examples{
if(interactive()){
s3_client_setup()
}
}
