% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_auth.R
\name{s3_connection_HL}
\alias{s3_connection_HL}
\title{S3 initialisation}
\usage{
s3_connection_HL(
  key = NA,
  secret = NA,
  main_folder = NA,
  bucket = NA,
  endpoint = NA,
  region = NA,
  config_path = TRUE
)
}
\arguments{
\item{key}{(chr) S3 key}

\item{secret}{(chr) s3 secret}

\item{main_folder}{(chr) folder added behind object}

\item{bucket}{(chr) s3 bucket}

\item{endpoint}{(chr) endpoint nécessaire pour accéder à S3}

\item{region}{(chr) s3 region, OVH == ""}

\item{config_path}{TRUE,FALSE or write location ex: "inst/config/"}
}
\value{
Rien, mais crée des variables environnement simplifiant la connexion à S3
}
\description{
S3 initialisation
}
\examples{
if(interactive()){
s3_connection_HL()
}
}
