% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_manipulation.R
\name{s3delete_HL}
\alias{s3delete_HL}
\title{Supprimer un fichier sur S3 (OVH ou AWS) avec aws.s3}
\usage{
s3delete_HL(
  object,
  bucket = NA,
  main_folder = TRUE,
  key = NA,
  secret = NA,
  endpoint = NA,
  region = NA
)
}
\arguments{
\item{object}{(chr) Fichier à supprimer (avec sous-dossier)}

\item{bucket}{(chr) Bucket S3}

\item{main_folder}{(log|chr) TRUE = utilise HL_S3_MAIN_FOLDER, sinon dossier ou ""}

\item{key}{(chr) S3 key}

\item{secret}{(chr) S3 secret}

\item{endpoint}{(chr) Endpoint S3}

\item{region}{(chr) Région}
}
\value{
TRUE si suppression OK, FALSE sinon
}
\description{
Supprimer un fichier sur S3 (OVH ou AWS) avec aws.s3
}
