% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_manipulation.R
\name{s3exist_HL}
\alias{s3exist_HL}
\title{Vérifier si un objet existe sur S3 (aws.s3, wrapper HL)}
\usage{
s3exist_HL(
  object,
  bucket = NA,
  main_folder = TRUE,
  key = NA,
  secret = NA,
  endpoint = NA,
  region = NA
)
}
\arguments{
\item{object}{Chemin de l'objet (sans main_folder)}

\item{bucket}{Bucket S3}

\item{main_folder}{(log) TRUE pour préfixer object par HL_S3_MAIN_FOLDER}

\item{key}{Key pour accès au S3}

\item{secret}{Secret pour accès au S3}

\item{endpoint}{Endpoint (base_url) pour accéder au S3 (ex: "s3.bhs.cloud.ovh.net")}

\item{region}{Region (parfois nécessaire). Pour OVH, peut être "".}
}
\value{
TRUE si l'objet existe, FALSE sinon
}
\description{
Vérifier si un objet existe sur S3 (aws.s3, wrapper HL)
}
