% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_manipulation.R
\name{s3list_HL}
\alias{s3list_HL}
\title{Liste les objets dans un bucket ou sous-dossier S3}
\usage{
s3list_HL(
  prefix = "",
  bucket = NA,
  main_folder = TRUE,
  key = NA,
  secret = NA,
  endpoint = NA,
  region = NA
)
}
\arguments{
\item{prefix}{Sous-dossier (peut être "")}

\item{bucket}{Nom du bucket}

\item{main_folder}{Si TRUE, préfixe automatiquement "HL_S3_MAIN_FOLDER"}

\item{key}{S3 key}

\item{secret}{S3 secret}

\item{endpoint}{Base URL S3}

\item{region}{Region (AWS requis, OVH="" )}
}
\value{
Liste d'objets (comme aws.s3::get_bucket)
}
\description{
Liste les objets dans un bucket ou sous-dossier S3
}
