% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_manipulation.R
\name{s3listdf_HL}
\alias{s3listdf_HL}
\title{Liste de fichier}
\usage{
s3listdf_HL(
  prefix = "",
  bucket = NA,
  main_folder = TRUE,
  key = NA,
  secret = NA,
  endpoint = NA,
  region = NA,
  max = NULL,
  type = ".rds",
  lastModified = "lubridate",
  ...
)
}
\arguments{
\item{prefix}{Sous-dossier (peut être "")}

\item{bucket}{Nom du bucket}

\item{main_folder}{Si TRUE, préfixe automatiquement "HL_S3_MAIN_FOLDER"}

\item{key}{S3 key}

\item{secret}{S3 secret}

\item{endpoint}{Base URL S3}

\item{region}{Region (AWS requis, OVH="" )}

\item{max}{maximum de fichiers retourner, max = INF pour tout avoir}

\item{type}{Pour ne filtrer qu'un extension de fichier}

\item{lastModified}{"lubridate"}

\item{...}{Autres paramètres transmis à une fonction interne ou à une méthode}
}
\value{
un dataframe avec les informations du dossier demandé
}
\description{
Retourne la liste des fichiers d'un bucket et dossier précis de S3
}
\examples{
if (interactive()){
files_all <- s3listdf_HL("OTP/legs_history")
}
}
