% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_manipulation.R
\name{s3readRDS_HL}
\alias{s3readRDS_HL}
\title{Fonction pour lire un .rds sur S3}
\usage{
s3readRDS_HL(
  object,
  bucket = NA,
  main_folder = NA,
  key = NA,
  secret = NA,
  endpoint = NA,
  region = NA
)
}
\arguments{
\item{object}{Fichier à lire (avec sous dossier)}

\item{bucket}{Bucket S3}

\item{main_folder}{Dossier à ajouter automatiquement devant l'objet}

\item{key}{Key pour accès au S3}

\item{secret}{Secret pour accès au S3}

\item{endpoint}{Enpoint pour accéder au S3}

\item{region}{Region (parfoit nécessaire), AWS est requis, OVH doit == ""}
}
\value{
Valeur du fichier .rds
}
\description{
Fonction pour lire un .rds sur S3
}
\examples{
if(interactive()){
s3readRDS_HL(object)
}
}
