% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_manipulation.R
\name{s3readRDS_HL_paws}
\alias{s3readRDS_HL_paws}
\title{Lire un fichier .rds depuis S3 avec paws}
\usage{
s3readRDS_HL_paws(
  object,
  bucket = NA,
  main_folder = NA,
  key = NA,
  secret = NA,
  endpoint = NA,
  region = NA
)
}
\arguments{
\item{object}{Fichier à lire (avec sous dossier)}

\item{bucket}{Bucket S3}

\item{main_folder}{Dossier à ajouter automatiquement devant l'objet}

\item{key}{Key pour accès au S3}

\item{secret}{Secret pour accès au S3}

\item{endpoint}{Endpoint pour accéder au S3}

\item{region}{Region (AWS requis, OVH peut être "")}
}
\value{
Valeur du fichier .rds
}
\description{
Lire un fichier .rds depuis S3 avec paws
}
\examples{
if(interactive()){
  s3readRDS_HL_paws("mon_fichier.rds")
}
}
