% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_manipulation.R
\name{s3saveRDS_HL}
\alias{s3saveRDS_HL}
\title{Fonction pour sauvegarder un .rds sur S3 (OVH compatible)}
\usage{
s3saveRDS_HL(
  value,
  object_name,
  bucket = NA,
  main_folder = TRUE,
  key = NA,
  secret = NA,
  endpoint = NA,
  region = NA
)
}
\arguments{
\item{value}{Objet R à sauvegarder}

\item{object_name}{Nom du fichier à créer dans S3 (peut inclure sous-dossiers)}

\item{bucket}{Bucket S3}

\item{main_folder}{(logique) Ajouter automatiquement HL_S3_MAIN_FOLDER ?}

\item{key}{S3 key}

\item{secret}{S3 secret}

\item{endpoint}{Endpoint OVH}

\item{region}{Région ("" pour OVH)}
}
\value{
TRUE si succès, erreur sinon
}
\description{
Fonction pour sauvegarder un .rds sur S3 (OVH compatible)
}
