% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_auth.R
\name{set_config_s3_access}
\alias{set_config_s3_access}
\title{Save S3 access}
\usage{
set_config_s3_access(
  s3_ACCESS_KEY_ID,
  s3_SECRET_ACCESS_KEY,
  s3_REGION,
  s3_ENDPOINT = "s3.amazonaws.com",
  config_path = "inst/app/data/"
)
}
\arguments{
\item{s3_ACCESS_KEY_ID}{AWS ACCESS KEY ID}

\item{s3_SECRET_ACCESS_KEY}{AWS SECRET ACCESS KEY}

\item{s3_REGION}{REGION}

\item{s3_ENDPOINT}{Endpoint (à spécifier si autre que AWS/S3)}

\item{config_path}{emplacement pour écrire un fichier de config local}
}
\value{
Rien. La fonction enregistre un fichier sur le disque.
}
\description{
Sert à enregistrer les informations nécessaires pour accéder à S3/AWS
}
\examples{
if (interactive()) {
  s3_ACCESS_KEY_ID <- "access_key"
  s3_SECRET_ACCESS_KEY <- "secret_key"
  s3_REGION <- "where"
  s3_ENDPOINT <- "end_point"
  set_config_s3_access("access_key", "secret_key", "region","end_point","s3_ENDPOINT")
}
}
