% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_auth.R
\name{set_config_s3_location}
\alias{set_config_s3_location}
\title{Set S3 bucket and main folder for this app}
\usage{
set_config_s3_location(
  s3_bucket,
  s3_main_folder,
  secure_key,
  config_path = "inst/app/data/"
)
}
\arguments{
\item{s3_bucket}{Nom de mon bucket S3}

\item{s3_main_folder}{Nom du dossier principal dans le bucket}

\item{secure_key}{Sel pour cripter les informations sensibles}

\item{config_path}{emplacement pour écrire un fichier de config local}
}
\value{
Rien. La fonction enregistre un fichier sur le disque.
}
\description{
Cette fonction sert à enregistrer un fichier contenant le bucket et dossier principal lié avec ce projet
}
\examples{
if (interactive()) {
  s3_bucket <- "my-bucket"
  s3_main_folder <- "my-folder"
  secure_key <- "123456"
  set_config_s3_location(s3_bucket, s3_main_folder, secure_key)
}
}
